FIXED: force cast into a uint when returning pthread_self
git-svn-id: svn://svn.code.sf.net/p/nel/code/trunk@341 bf685bcc-7839-4fcd-9254-36823579bede

+++ nel/src/misc/common.cpp
@@ -115,7 +115,7 @@ uint getThreadId()
 #ifdef NL_OS_WINDOWS
 	return GetCurrentThreadId();
 #elif defined NL_OS_UNIX
-	return pthread_self();
+	  return uint(pthread_self());
 	// doesnt work on linux kernel 2.6	return getpid();
 #endif
